% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel_functions.R
\name{.check_dd}
\alias{.check_dd}
\title{Check data dictionary (generic)}
\usage{
.check_dd(dd, ds = NULL)
}
\arguments{
\item{dd}{Data dictionary object}

\item{ds}{Corresponding dataset object}
}
\description{
Check data dictionary (generic)
}
\details{
Reports errors or issues as warnings.
}
