% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_sattr}
\alias{check_sattr}
\title{Check sample attributes file}
\usage{
check_sattr(dsfile, ddfile = NULL, na_vals = c("NA", "N/A", "na", "n/a"),
  samp_exp = NULL, sampleID_col = "SAMPLE_ID", topmed = FALSE)
}
\arguments{
\item{dsfile}{Path to the data file on disk}

\item{ddfile}{Path to the data dictionary file on disk}

\item{na_vals}{Vector of strings that should be read in as NA/missing in data file (see details of \code{.read_ds_file})}

\item{samp_exp}{List of expected sample IDs}

\item{sampleID_col}{Column name for sample-level ID}

\item{topmed}{Logical to indicate TOPMed study}
}
\value{
satt_report, a list of the following issues (when present):
\item{missing_vars}{Missing and required variables}
\item{dup_samples}{List of duplicated sample IDs}
\item{blank_idx}{Row index of blank/missing sample IDs}
\item{dd_errors}{Differences in fields between data file and data dictionary}
\item{extra_samples}{Samples in data file missing from \code{ssm_exp}}
\item{missing_samples}{Samples in \code{ssm_exp} missing from data file}
\item{missing_topmed_vars}{Missing and required variables for TOPMed}
}
\description{
Check contents of a sample attributes file for dbGaP posting.
}
\details{
The sample attributes file should be a tab-delimited .txt file.
When (\code{topmed = TRUE}) checks presence of additional, TOPMed-specific
sample attributes variables: SEQUENCING_CENTER, Funding_Source, TOPMed_Phase, 
TOPMed_Project, Study_Name.

If a data dictionary is provided (\code{ddfile != NULL}), additionally checks 
correspondence between column names in data file and entries in data dictionary.
Data dictionary files can be Excel (.xls, .xlsx) or tab-delimited .txt.
}
