% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel_functions.R
\name{.read_ds_file}
\alias{.read_ds_file}
\title{Read in a data file}
\usage{
.read_ds_file(filename, dd = FALSE)
}
\arguments{
\item{filename}{The path to the file on disk}

\item{dd}{Logical, where \code{TRUE} indicates a data dictionary file}
}
\value{
A data frame from the file
}
\description{
Works for tab-delimited (.txt) data files
}
\details{
The only string considered to be NA is the blank string: "". "NA" does not necessarily indicate
NA: for example, it could be an encoded value meaning "North America".

dbGaP dataset files should have column headers as the first row. If the input violates this, e.g. additional header rows are present, a warning is returned but the file is still read in.
}
