% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.check_dd}
\alias{.check_dd}
\title{Check data dictionary (generic)}
\usage{
.check_dd(dd, ds = NULL)
}
\arguments{
\item{dd}{Data dictionary (DD) object}

\item{ds}{Corresponding dataset (DS) object}
}
\value{
dd_report, a list of the following issues (when present):
\item{lowercase}{Logical flag indicating non-upper case variable names}
\item{varname_vardesc}{Logical flag indicating first two variables are not VARNAME, VARDESC}
\item{missing_vars}{Missing and required variables}
\item{extra_vars}{Extra variables}
\item{vals_warnings}{Vector of warnings about VALUES columns}
\item{missing_dsvars}{Variables present in DS but not defined in DD}
\item{min_errs}{Variables for which DS value are < DD MIN}
\item{max_errs}{Variables for which DS value are > DD MAX}
\item{illegal_vars}{Variable names containing illegal characters: '\', '/', ',' (comma), or 'dbGaP' are present}
}
\description{
Check data dictionary (generic)
}
\details{
Reports errors or issues with DD file.
When the corresponding DS file is also provided, checks for consistency between the two.
}
